moved back fetch to each component and styled PopulaList page
